整理：voice_pipeline/mod.rs 测试下沉到子文件 + test_doubles.rs
mod.rs 477 行中 422 行是测试，与 voice_pipeline/ 子目录拆分
目的相悖。按测试对象下沉到 builder.rs/handlers.rs/context.rs/
sink.rs 各自模块的 #[cfg(test)] mod tests；共享 fixture
（FakeListener/FakeRecorder/MockSink/FakeOutput 等）下到 test_doubles.rs
供兄弟测试模块 use。

mod.rs 缩到 51 行，只保留装配。

- builder.rs tests: 5（原 4 + run() 入口错误上报测试，故障点 build_context）
- context.rs tests: 2
- handlers.rs tests: 9（含 4 个 select_text + 3 个 process_transcription_result + 2 个 handle_*_record）
- test_doubles.rs: 收 FakeListener/FakeRecorder/MockSink/FakeOutput，pub(super) 暴露

测试用例总数：16 → 16（无变化）。
